+++ html/layouts/joomla/edit/frontediting_modules.php
@@ -1,9 +1,8 @@
 <?php
 /**
- * @package     Joomla.Site
- * @subpackage  Layout
+ * @package     Lightning
  *
- * @copyright   Copyright (C) 2005 - 2019 Open Source Matters, Inc. All rights reserved.
+ * @copyright   Copyright (C) 2020 JoomJunk. All rights reserved.
  * @license     GNU General Public License version 2 or later; see LICENSE.txt
  */
 
@@ -68,5 +67,5 @@ $moduleHtml = preg_replace(
 if ($count)
 {
 	HTMLHelper::_('stylesheet', 'legacy/frontediting.css', ['version' => 'auto', 'relative' => true]);
-	HTMLHelper::_('script', 'legacy/frontediting.min.js', ['version' => 'auto', 'relative' => true]);
+	HTMLHelper::_('script', 'legacy/frontediting.min.js', ['version' => 'auto', 'relative' => true], ['type' => 'module']);
 }