+++ index.php
@@ -13,27 +13,22 @@ use Joomla\CMS\HTML\HTMLHelper;
 use Joomla\CMS\Language\Text;
 use Joomla\CMS\Uri\Uri;
 
+include_once __DIR__ . '/helper/metas.php';
+include_once __DIR__ . '/helper/styles.php';
+include_once __DIR__ . '/helper/scripts.php';
+
 /** @var JDocumentHtml $this */
 
-$app  = Factory::getApplication();
-$lang = $app->getLanguage();
-
-// Detecting Active Variables
-$option   = $app->input->getCmd('option', '');
-$view     = $app->input->getCmd('view', '');
-$layout   = $app->input->getCmd('layout', '');
-$task     = $app->input->getCmd('task', '');
-$itemid   = $app->input->getCmd('Itemid', '');
-$sitename = htmlspecialchars($app->get('sitename'), ENT_QUOTES, 'UTF-8');
-$menu     = $app->getMenu()->getActive();
-$pageclass = $menu->getParams()->get('pageclass_sfx');
+$app           = Factory::getApplication();
+$sitename      = htmlspecialchars($app->get('sitename'), ENT_QUOTES, 'UTF-8');
+$pageclass     = $app->getMenu()->getActive()->getParams()->get('pageclass_sfx');
+$themeSwitcher = (boolean)$this->params->get('theme-switcher', true);
 
 // Template params
-$themeSwitcher = (boolean)$this->params->get('theme-switcher', true);
 if ($themeSwitcher)
 {
 	HTMLHelper::_('stylesheet', 'switch.css', ['version' => 'auto', 'relative' => true]);
-	HTMLHelper::_('script', 'switch.min.js', ['version' => 'auto', 'relative' => true]);
+	HTMLHelper::_('script', 'switch.min.js', ['version' => 'auto', 'relative' => true], ['type' => 'module']);
 }
 
 // Font Awesome
@@ -68,32 +63,50 @@ if ($this->countModules('sidebar-right'))
 	$hasSidebar .= ' has-sidebar-right';
 }
 
-$hasBanner = '';
-
-if ($this->countModules('banner'))
-{
-	$hasBanner = 'has-banner';
-}
+$faviconPath = 'templates/' . $this->template . '/favicon';
 
 $this->setMetaData('viewport', 'width=device-width, initial-scale=1');
+$this->setMetaData('theme-color', '#ffffff');
+$this->setMetaData('msapplication-config', $faviconPath . '/browserconfig.xml');
+$this->setMetaData('msapplication-TileColor', '#ffffff');
+
+$this->addCustomTag('<link rel="apple-touch-icon" sizes="180x180" href="' . $faviconPath . '/apple-touch-icon.png">');
+$this->addCustomTag('<link rel="icon" type="image/png" sizes="32x32" href="' . $faviconPath . '/favicon-32x32.png">');
+$this->addCustomTag('<link rel="icon" type="image/png" sizes="16x16" href="' . $faviconPath . '/favicon-16x16.png">');
+$this->addCustomTag('<link rel="manifest" href="' . $faviconPath . '/site.webmanifest">');
+$this->addCustomTag('<link rel="mask-icon" href="' . $faviconPath . '/safari-pinned-tab.svg" color="#5bbad5">');
+$this->addCustomTag('<link rel="shortcut icon" href="' . $faviconPath . '/favicon.ico">');
+
+$menu         = $this->getBuffer('modules', 'menu', $attribs = ['style' => 'none']);
+$search       = $this->getBuffer('modules', 'search', $attribs = ['style' => 'none']);
+$banner       = $this->getBuffer('modules', 'banner', $attribs = ['style' => 'default']);
+$topA         = $this->getBuffer('modules', 'top-a', $attribs = ['style' => 'default']);
+$topB         = $this->getBuffer('modules', 'top-b', $attribs = ['style' => 'default']);
+$sidebarLeft  = $this->getBuffer('modules', 'sidebar-left', $attribs = ['style' => 'default']);
+$mainTop      = $this->getBuffer('modules', 'main-top', $attribs = ['style' => 'default']);
+$message      = $this->getBuffer('message');
+$breadcrumbs  = $this->getBuffer('modules', 'breadcrumbs', $attribs = ['style' => 'none']);
+$component    = $this->getBuffer('component');
+$mainBottom   = $this->getBuffer('modules', 'main-bottom', $attribs = ['style' => 'default']);
+$sidebarRight = $this->getBuffer('modules', 'sidebar-right', $attribs = ['style' => 'default']);
+$bottomA      = $this->getBuffer('modules', 'bottom-a', $attribs = ['style' => 'default']);
+$bottomB      = $this->getBuffer('modules', 'bottom-b', $attribs = ['style' => 'default']);
+$footer       = $this->getBuffer('modules', 'footer', $attribs = ['style' => 'none']);
+$debug        = $this->getBuffer('modules', 'debug', $attribs = ['style' => 'none']);
+$metas        = $this->getBuffer('metas');
+$styles       = $this->getBuffer('styles');
+$scripts      = $this->getBuffer('scripts');
+
+$cachesStyleSheets = json_encode(array_values($styles));
 ?>
 <!DOCTYPE html>
 <html lang="<?php echo $this->language; ?>" dir="<?php echo $this->direction; ?>">
 <head>
-	<jdoc:include type="metas" />
+	<?php echo $metas; ?>
 	<style><?php echo $css; ?></style>
 </head>
-
-<body class="site-grid site <?php echo $option
-	. ' view-' . $view
-	. ($layout ? ' layout-' . $layout : ' no-layout')
-	. ($task ? ' task-' . $task : ' no-task')
-	. ($itemid ? ' itemid-' . $itemid : '')
-	. ' ' . $pageclass
-	. $hasSidebar;
-	echo ($this->direction == 'rtl' ? ' rtl' : '');
-?>">
-	<header class="grid-child container-header full-width header <?php echo $hasBanner; ?>">
+<body class="site-grid site <?php echo $pageclass . $hasSidebar; ?>">
+	<header class="grid-child container-header full-width header <?php echo $this->countModules('banner') ? 'has-banner' : ''; ?>">
 		<nav class="navbar">
 			<div class="navbar-brand">
 				<a href="<?php echo $this->baseurl; ?>/">
@@ -107,10 +120,10 @@ $this->setMetaData('viewport', 'width=device-width, initial-scale=1');
 
 			<?php if ($this->countModules('menu') || $this->countModules('search')) : ?>
 				<div class="navbar-menu">
-					<jdoc:include type="modules" name="menu" style="none" />
+					<?php echo $menu; ?>
 					<?php if ($this->countModules('search')) : ?>
 						<div>
-							<jdoc:include type="modules" name="search" style="none" />
+							<?php echo $search; ?>
 						</div>
 					<?php endif; ?>
 				</div>
@@ -127,74 +140,78 @@ $this->setMetaData('viewport', 'width=device-width, initial-scale=1');
 
 	<?php if ($this->countModules('banner')) : ?>
 	<div class="grid-child full-width container-banner">
-		<jdoc:include type="modules" name="banner" style="default" />
+		<?php echo $banner; ?>
 	</div>
 	<?php endif; ?>
 
 	<?php if ($this->countModules('top-a')) : ?>
 	<div class="grid-child container-top-a">
-		<jdoc:include type="modules" name="top-a" style="default" />
+		<?php echo $topA; ?>
 	</div>
 	<?php endif; ?>
 
 	<?php if ($this->countModules('top-b')) : ?>
 	<div class="grid-child container-top-b">
-		<jdoc:include type="modules" name="top-b" style="default" />
+		<?php echo $topB; ?>
 	</div>
 	<?php endif; ?>
 
 	<?php if ($this->countModules('sidebar-left')) : ?>
 	<div class="grid-child container-sidebar-left">
-		<jdoc:include type="modules" name="sidebar-left" style="default" />
+		<?php echo $sidebarLeft; ?>
 	</div>
 	<?php endif; ?>
 
 	<div class="grid-child container-component">
-		<jdoc:include type="modules" name="main-top" style="default" />
-		<jdoc:include type="message" />
-		<jdoc:include type="modules" name="breadcrumbs" style="none" />
-		<jdoc:include type="component" />
-		<jdoc:include type="modules" name="main-bottom" style="default" />
+		<?php echo $mainTop; ?>
+		<?php echo $message; ?>
+		<?php echo $breadcrumbs; ?>
+		<?php echo $component; ?>
+		<?php echo $mainBottom; ?>
 	</div>
 
 	<?php if ($this->countModules('sidebar-right')) : ?>
 	<div class="grid-child container-sidebar-right">
-		<jdoc:include type="modules" name="sidebar-right" style="default" />
+		<?php echo $sidebarRight; ?>
 	</div>
 	<?php endif; ?>
 
 	<?php if ($this->countModules('bottom-a')) : ?>
 	<div class="grid-child container-bottom-a">
-		<jdoc:include type="modules" name="bottom-a" style="default" />
+		<?php echo $bottomA; ?>
 	</div>
 	<?php endif; ?>
 
 	<?php if ($this->countModules('bottom-b')) : ?>
 	<div class="grid-child container-bottom-b">
-		<jdoc:include type="modules" name="bottom-b" style="default" />
+		<?php echo $bottomB; ?>
 	</div>
 	<?php endif; ?>
 
 	<?php if ($this->countModules('footer')) : ?>
 	<footer class="grid-child container-footer full-width footer">
 		<div class="container">
-			<jdoc:include type="modules" name="footer" style="none" />
+			<?php echo $footer; ?>
 		</div>
 	</footer>
 	<?php endif; ?>
 
-	<jdoc:include type="modules" name="debug" style="none" />
+	<?php echo $debug; ?>
 
 	<?php if ($themeSwitcher) : ?>
 	<script>
 		(() => {
 			const theme = localStorage.getItem('theme') ?? 'is-light'
 			document.documentElement.classList.add(theme)
+
+			const styles = <?php echo $cachesStyleSheets; ?>;
+			styles.forEach(item => {
+				document.body.insertAdjacentHTML('beforeend', item);
+			})
 		})()
 	</script>
 	<?php endif; ?>
 
-	<jdoc:include type="styles" />
-	<jdoc:include type="scripts" />
+	<?php echo $scripts; ?>
 </body>
 </html>